+++ components/styles/composite/Accordion/index.ts
@@ -3,5 +3,3 @@ SPDX-FileCopyrightText: 2026 Humdek, University of Bern
 SPDX-License-Identifier: MPL-2.0
 */
 export { Accordion } from './Accordion';
-export { AccordionContext, useAccordionContext, useAccordionState } from './Accordion.hooks';
-export type { IAccordionContext } from './Accordion.types';